requirements.txt
alembic==1.18.4
blinker==1.9.0
certifi==2026.2.25
click==8.3.1
cloudinary==1.44.1
colorama==0.4.6
dnspython==2.8.0
email-validator==2.3.0
Flask==3.1.3
Flask-Login==0.6.3
Flask-Migrate==4.1.0
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.2
Flask-Limiter==3.5.0
Flask-Mail==0.9.1
Flask-Cors==4.0.0
greenlet==3.3.2
gunicorn==25.1.0
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
Mako==1.3.10
MarkupSafe==3.0.3
packaging==26.0
psycopg2-binary==2.9.11
python-dotenv==1.2.2
pyspellchecker==0.7.2
six==1.17.0
SQLAlchemy==2.0.48
typing_extensions==4.15.0
urllib3==2.6.3
Werkzeug==3.1.6
WTForms==3.2.1

# Development/Testing (uncomment when needed)
# pytest==7.4.3
# pytest-flask==1.3.0
# pytest-cov==4.1.0
